refactored the folder and file structure for type initialization and definition (just moved code around, no further changes)
