+++ packages/typir/src/features/operator.ts
@@ -7,7 +7,7 @@
 import { Type } from '../graph/type-node.js';
 import { FunctionKind, FunctionKindName, isFunctionKind, NO_PARAMETER_NAME } from '../kinds/function-kind.js';
 import { TypirServices } from '../typir.js';
-import { TypeInitializer } from '../utils/type-initialization.js';
+import { TypeInitializer } from '../initialization/type-initializer.js';
 import { NameTypePair, TypeInitializers } from '../utils/utils-definitions.js';
 import { toArray } from '../utils/utils.js';
