+++ packages/typir/src/index.ts
@@ -17,6 +17,10 @@ export * from './features/validation.js';
 export * from './graph/type-edge.js';
 export * from './graph/type-graph.js';
 export * from './graph/type-node.js';
+export * from './initialization/type-initializer.js';
+export * from './initialization/type-reference.js';
+export * from './initialization/type-selector.js';
+export * from './initialization/type-waiting.js';
 export * from './kinds/class-kind.js';
 export * from './kinds/fixed-parameters-kind.js';
 export * from './kinds/function-kind.js';
@@ -27,6 +31,5 @@ export * from './kinds/top-kind.js';
 export * from './utils/dependency-injection.js';
 export * from './utils/test-utils.js';
 export * from './utils/utils.js';
-export * from './utils/type-initialization.js';
 export * from './utils/utils-definitions.js';
 export * from './utils/utils-type-comparison.js';
